+++ package.json
@@ -33,17 +33,16 @@
     }
   },
   "dependencies": {
-    "classnames": "^2.2.6",
-    "gatsby": "^2.14.0",
-    "gatsby-plugin-google-analytics": "^2.1.11",
-    "gatsby-plugin-manifest": "^2.2.9",
-    "gatsby-plugin-offline": "^2.2.10",
-    "gatsby-plugin-react-helmet": "^3.1.5",
-    "gatsby-plugin-sharp": "^2.2.18",
-    "gatsby-plugin-typography": "^2.3.5",
-    "gatsby-source-filesystem": "^2.1.18",
-    "gatsby-transformer-json": "^2.2.6",
-    "gatsby-transformer-sharp": "^2.2.12",
+    "gatsby": "^2.15.21",
+    "gatsby-plugin-google-analytics": "^2.1.17",
+    "gatsby-plugin-manifest": "^2.2.18",
+    "gatsby-plugin-offline": "^3.0.8",
+    "gatsby-plugin-react-helmet": "^3.1.8",
+    "gatsby-plugin-sharp": "^2.2.25",
+    "gatsby-plugin-typography": "^2.3.8",
+    "gatsby-source-filesystem": "^2.1.26",
+    "gatsby-transformer-json": "^2.2.9",
+    "gatsby-transformer-sharp": "^2.2.16",
     "react": "^16.9.0",
     "react-dom": "^16.9.0",
     "react-helmet": "^5.2.1",
